chore(RHTAPWATCH-1182): create secret for local registry
The dockerconfigjson file that is used by some of the pipelines to
hold info for authentication with the registry is still needed even
though the registry is not using any authentication.

With this change, we introduce a dummy secret that allows this file to
be created without any authentication info.

Signed-off-by: Yftach Herzog <yherzog@redhat.com>